remove validation code again, lets reuse what we have and refactor that later
