+++ apps/tailwind-components/plugins/onFirstView.ts
@@ -1,9 +1,9 @@
-// plugins/lazy-load-directive.client.ts
 import { type DirectiveBinding } from 'vue'
 
 type LazyLoadBindingValue = () => Promise<void> | void
 
 export default defineNuxtPlugin((nuxtApp) => {
+    //will call the binded function when the viewport is getting close to view the element
     nuxtApp.vueApp.directive('on-first-view', {
         mounted(el: HTMLElement, binding: DirectiveBinding<LazyLoadBindingValue>) {
             const options: IntersectionObserverInit = {
