+++ apps/tailwind-components/utils/flattenObject.ts
@@ -1,0 +1,18 @@
+export function flattenObject(object: Record<string, any>): string {
+  if (typeof object === "object") {
+    let result = "";
+    Object.keys(object).forEach((key) => {
+      if (object[key] === null) {
+        return;
+      }
+      if (typeof object[key] === "object") {
+        result += flattenObject(object[key]);
+      } else {
+        result += " " + object[key];
+      }
+    });
+    return result;
+  } else {
+    return object;
+  }
+}
